fix convert exceptions, and rename some conflictive packages
