+++ app/src/hnqis/res/values/strings.xml
@@ -258,7 +258,6 @@ evaluation"</string>
 <string name="get_db_file_dialog">Please select a valid db file</string>
 <string name="start">Start</string>
 <string name="on">"on"</string>
-<string name="date_month_text_format">"MMMM d',' YYYY"</string>
 
 
 <string name="add">"Add"</string>
