+++ app/src/main/java/org/eyeseetea/malariacare/data/database/utils/monitor/MonitorMessagesBuilder.java
@@ -24,6 +24,7 @@ import android.util.Log;
 import android.webkit.WebView;
 
 import org.eyeseetea.malariacare.R;
+import org.eyeseetea.malariacare.data.database.utils.PreferencesState;
 
 /**
  * Created by arrizabalaga on 26/07/16.
@@ -33,18 +34,6 @@ public class MonitorMessagesBuilder {
     private static final String TAG="MonitorMessagesBuilder";
     public static final String JSON_MAP = "{'assesmentUnderTaken':'%s','target': '%s','qualityOfCare': '%s','months':'%s'}";
     public static final String JAVASCRIPT_UPDATE_TABLE = "javascript:initContext(%s)";
-
-    /**
-     * Required to inyect title according to current language
-     */
-    private Context context;
-
-    /**
-     * Default constructor
-     */
-    public MonitorMessagesBuilder(Context context) {
-        this.context = context;
-    }
 
     /**
      * Adds calculated entries to the given webView
@@ -58,10 +47,10 @@ public class MonitorMessagesBuilder {
     }
 
     private String buildJSON() {
-        String assesmentUnderTaken = context.getResources().getString(R.string.monitor_js_assessments);
-        String target = context.getResources().getString(R.string.monitor_js_target);
-        String qualityOfCare = context.getResources().getString(R.string.monitor_js_quality_of_care);
-        String months = context.getResources().getString(R.string.monitor_js_months);
+        String assesmentUnderTaken = PreferencesState.getInstance().getContext().getResources().getString(R.string.monitor_js_assessments);
+        String target = PreferencesState.getInstance().getContext().getResources().getString(R.string.monitor_js_target);
+        String qualityOfCare = PreferencesState.getInstance().getContext().getResources().getString(R.string.monitor_js_quality_of_care);
+        String months = PreferencesState.getInstance().getContext().getResources().getString(R.string.monitor_js_months);
         return String.format(JSON_MAP,assesmentUnderTaken,target,qualityOfCare,months);
     }
