+++ app/src/main/java/org/eyeseetea/malariacare/data/database/utils/monitor/facilities/FacilityTableBuilderByOrgUnit.java
@@ -17,7 +17,7 @@
  *  along with Foobar.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package org.eyeseetea.malariacare.data.database.utils.monitor.facility;
+package org.eyeseetea.malariacare.data.database.utils.monitor.facilities;
 
 import android.content.Context;
 import android.util.Log;
@@ -42,10 +42,9 @@ public class FacilityTableBuilderByOrgUnit extends  FacilityTableBuilderBase {
      * Default constructor
      *
      * @param surveys
-     * @param context
      */
-    public FacilityTableBuilderByOrgUnit(List<SurveyDB> surveys, Context context) {
-        super(surveys, context);
+    public FacilityTableBuilderByOrgUnit(List<SurveyDB> surveys) {
+        super(surveys);
         this.facilityTableDataMap = new HashMap<>();
     }
