+++ app/src/main/java/org/eyeseetea/malariacare/data/database/utils/monitor/pies/PieBuilderByProgram.java
@@ -17,14 +17,14 @@
  *  along with Foobar.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package org.eyeseetea.malariacare.data.database.utils.monitor.pie;
+package org.eyeseetea.malariacare.data.database.utils.monitor.pies;
 
-import android.content.Context;
 import android.webkit.WebView;
 
 import org.eyeseetea.malariacare.R;
-import org.eyeseetea.malariacare.data.database.model.OrgUnitDB;
+import org.eyeseetea.malariacare.data.database.model.ProgramDB;
 import org.eyeseetea.malariacare.data.database.model.SurveyDB;
+import org.eyeseetea.malariacare.data.database.utils.PreferencesState;
 
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -34,30 +34,30 @@ import java.util.Map;
 /**
  * Created by idelcano on 23/08/2016.
  */
-public class PieBuilderByOrgUnit extends PieBuilderBase {
-    public static final String JAVASCRIPT_UPDATE_CHARTS = "javascript:setOrgUnitPieData(%s)";
+public class PieBuilderByProgram extends PieBuilderBase {
+    public static final String JAVASCRIPT_UPDATE_CHARTS = "javascript:setProgramPieData(%s)";
 
     /**
      * Map of entries per program
      */
-    private Map<OrgUnitDB,PieDataByOrgUnit> pieTabGroupDataMap;
+    private Map<ProgramDB,PieDataByProgram> pieTabGroupDataMap;
     /**
      * Default constructor
      *
      * @param surveys
-     * @param context
      */
-    public PieBuilderByOrgUnit(List<SurveyDB> surveys, Context context) {
-        super(surveys, context);
-        pieTabGroupDataMap=new HashMap<>();
+    public PieBuilderByProgram(List<SurveyDB> surveys) {
+        super(surveys);
+        pieTabGroupDataMap = new HashMap<>();
     }
+
     /**
      * Adds calculated entries to the given webView
      * @param webView
      */
     public void addDataInChart(WebView webView){
         //Build entries
-        List<PieDataByOrgUnit> entries=build(surveys);
+        List<PieDataByProgram> entries=build(surveys);
         //Inyect entries in view
         injectDataInChart(webView, entries);
         buildJSONArray(entries);
@@ -65,41 +65,38 @@ public class PieBuilderByOrgUnit extends PieBuilderBase {
     }
     private void build(SurveyDB survey) {
         //Get the program
-        OrgUnitDB orgUnit=survey.getOrgUnit();
+        ProgramDB program=survey.getProgram();
 
         //Get the entry for that program
-        PieDataByOrgUnit pieTabGroupData = pieTabGroupDataMap.get(orgUnit);
+        PieDataByProgram pieTabGroupData = pieTabGroupDataMap.get(program);
 
         //First time no entry
         if(pieTabGroupData ==null){
-            pieTabGroupData =new PieDataByOrgUnit(orgUnit);
-            pieTabGroupDataMap.put(orgUnit, pieTabGroupData);
+            pieTabGroupData =new PieDataByProgram(program);
+            pieTabGroupDataMap.put(program, pieTabGroupData);
         }
         //Increment surveys for that month
         pieTabGroupData.incCounter(survey.getMainScore());
     }
-
-
-    private List<PieDataByOrgUnit> build(List<SurveyDB> surveys) {
+    private List<PieDataByProgram> build(List<SurveyDB> surveys) {
         for(SurveyDB survey:surveys){
             build(survey);
         }
 
         return new ArrayList(pieTabGroupDataMap.values());
     }
-    private void injectDataInChart(WebView webView, List<PieDataByOrgUnit> entries) {
+    private void injectDataInChart(WebView webView, List<PieDataByProgram> entries) {
         //Build array JSON
         String json=buildJSONArray(entries);
 
         //Inyect in browser
         inyectInBrowser(webView, JAVASCRIPT_UPDATE_CHARTS, json);
     }
-
-    private String buildJSONArray(List<PieDataByOrgUnit> entries){
+    private String buildJSONArray(List<PieDataByProgram> entries){
         String arrayJSON="[";
         int i=0;
-        for(PieDataByOrgUnit pieTabGroupData :entries){
-            String pieJSON= pieTabGroupData.toJSON(context.getString(R.string.dashboard_tip_pie_chart));
+        for(PieDataByProgram pieTabGroupData :entries){
+            String pieJSON= pieTabGroupData.toJSON(PreferencesState.getInstance().getContext().getString(R.string.dashboard_tip_pie_chart));
             arrayJSON+=pieJSON;
             i++;
             if(i!=entries.size()){
